Add/modular js (#72)
* Move Fractal into devDependencies
* Use modular JS
